Move the IP list to it's own file for clarity
