+++ lib/authorization/index.js
@@ -4,60 +4,14 @@ const _ = require('lodash');
 const jwt = require('jsonwebtoken');
 const shiroTrie = require('shiro-trie');
 
+const ipdelaylist = require('./delaylist')();
 const consts = require('./../constants');
-
-const log_green = '\x1B[32m';
-const log_red = '\x1b[31m';
-const log_reset = '\x1B[0m';
-const LOG_GRANTED = log_green + 'GRANTED: ' + log_reset;
-const LOG_DENIED = log_red + 'DENIED:  ' + log_reset;
-
-const ipDelayList = {};
-
-const DELAY_ON_FAIL = 5000;
-const FAIL_AGE = 60000;
 
 const sleep = require('util').promisify(setTimeout);
 
-function addFailedRequest (ip) {
-  let entry = ipDelayList[ip];
-  const now = Date.now();
-  if (!entry) {
-    ipDelayList[ip] = now + DELAY_ON_FAIL;
-    return;
-  }
-  if (now >= entry) { entry = now; }
-  ipDelayList[ip] = entry + DELAY_ON_FAIL;
-}
-
-function shouldDelayRequest (ip) {
-  let entry = ipDelayList[ip];
-  let now = Date.now();
-  if (entry) {
-    if (now < entry) {
-      return entry - now;
-    }
-  }
-  return false;
-}
-
-function requestSucceeded (ip) {
-  if (ipDelayList[ip]) {
-    delete ipDelayList[ip];
-  }
-}
-
-// Clear items older than a minute
-
-setTimeout(function clearList () {
-  for (var key in ipDelayList) {
-    if (ipDelayList.hasOwnProperty(key)) {
-      if (Date.now() > ipDelayList[key] + FAIL_AGE) {
-        delete ipDelayList[key];
-      }
-    }
-  }
-}, 30000);
+const addFailedRequest = ipdelaylist.addFailedRequest;
+const shouldDelayRequest = ipdelaylist.shouldDelayRequest;
+const requestSucceeded = ipdelaylist.requestSucceeded;
 
 function getRemoteIP (req) {
   return req.headers['x-forwarded-for'] || req.connection.remoteAddress;
